requirements.txt
numpy==1.15.4
scipy==1.0.0
progressbar2==3.38.0
python_speech_features==0.6
scikit-learn==0.19.1
